Simplify team section to image cards
